Readme (#212)
Let me summon

@telepath37 @khatril @lcd2yyz for awareness - this aims to lower the
entry bar to sell this more efficiently to our end user base and beyond.
Feel free to give feedback. You might find it easier to see from GH
preview (as opposed to PR changes):
https://github.com/insightsengineering/tlg-catalog/tree/readme

@cicdguy asking for overall review
@vedhav (only if you have bandwidth) - asking for review of
package/README.md file. You were somewhat included in the prior
discussion and I ask you to check if the instructions are correct and
clear.
And maybe @chlebowa (only if you have bandwidth) - for review of
package/README.md in particular. I assume you were not yet included in
the prior discussion and I want you to check whether the instructions
are clear from the perspective of the new developer.
For both of you, the expectation is a ~_basic_~ actually one level more
than _basic_ awareness of what's happening - not necessarily _complete_
understanding. This is not meant to be complete manual - just a README
file.

Closes https://github.com/insightsengineering/tlg-catalog/issues/111

---------

Signed-off-by: Aleksander Chlebowski <114988527+chlebowa@users.noreply.github.com>
Signed-off-by: Pawel Rucki <12943682+pawelru@users.noreply.github.com>
Co-authored-by: Aleksander Chlebowski <114988527+chlebowa@users.noreply.github.com>