move R CMD CHECK (stable) from docs to check (#236)
The rationale:
Each modification of an article should be tested against both stable and
devel in the feature branch before the merge. Currently this is
happening only after merge.

Key changes:
- moved `r-cmd-stable` from "docs" workflow into "check"
- `r-cmd-stable` will inherit triggers from that workflow and became
executed in the PR
- removed `r-cmd-stable-notification` as `r-cmd-stable` will be executed
in the PR
UPDATE: actually I think it's incorrect. We might also need
`r-cmd-stable` (alongside notifications) to be executed on a schedule.
- `publish-stable` job (which renders and publish catalog) is
unconditional to `r-cmd-stable`. This is good because `r-cmd-stable` is
covered already inside PR.

Small enhancements:
- remove `enforce-note-blocklist` with its `note-blocklist` because this
is a simple and dummy package

I was also thinking about renaming existing `r-cmd` into `r-cmd-devel`
but decided to keep it as is.

Please note this PR targets a feature branch and not main.