stable and devel snapshot testing (#235)
Fundamental changes:
- remove `-as-cran` to enable snapshot testing. Currently, we render the
whole book during the testing setup but we don't actually run the
snapshot tests. This is a buggy behaviour that needs to be fixed.
- in order to run both devel and stable tests, we need to have separate
snapshot files. Otherwise one of the profile would be most likely
failing because it would compare to the snapshot coming from the other
one.
- for this, we need to make sure to render the book during test setup
using appropriate profile. This is achieved by setting `QUARTO_PROFILE`
env var in the workflow definition. This will be then picked up by
Quarto itself.
- as a direct consequence - all the existing snapshots needs to be
recreated using both `stable` and `devel` profiles. Luckily, git
recognizes changes as file rename (instead of remove & add).
- Unfortunately, testthat does not support multi-levels directory name
to store snapshot file. It does not support custom names either. (I
created tickets - let's see what they say). Therefore we need to
_collapse_ this structure into a single string. I was thinking about the
naming convention. It would be great to have `{variant}/path/{id}.md`
but not possible :( I ran this using `{path}_{variant}/snaps.md` but I
am happy to change it to something else.

Enhancements:
- use full path of the article (including `{tables/listings/graphs} /
{subcategory}`, e.g. `tables/ADA/adat01.qmd` instead of `adat.qmd`) in
the `.Rds` file path (created by knitr hooks inside `testthat/_data`) as
well as snapshot file (`variant` argument of `testthat::test_snapshot()`
inside `_snaps` directory). Previously it was based on the basename of
an article only (e.g. `adat01`) which is not super robust - there might
be articles of the same basename in a separate directories.
- simplify the main testing function - see
`package/tests/testthat/test-snaps.R` file. I changed this to a simple
for loop through the articles. I think it's easier to follow and it
saves from redundant steps of e.g. removing ".rds" parts and adding this
back soon after.
- Currently we have two test files for markdown and graphs respectively.
This results in a different naming convention of resulting files.
Additionally, some parts are repeated. Last but not the least, graphs
tests are skipped for some reasons. I have combined this all into a
single `test-snaps.R` file assuming future work for graphs type of
tests.

Other comments / thoughts - just for a record:
- I have been thinking whether we actually need both stable and devel
snapshot testing from the perspective of what's the value. I believe
these are needed for potential updates of articles - a single article
must be tested against both profiles and a developer should be notified
whether changes results in output changes.
It's a little bit more complicated from the perspective of
reverse-dependency-checks from within e.g. tern. There, we need to limit
only to development profile most likely. But I left this problem for the
future
- Currently, we have a single test file `test-snaps.R` that does for
loop for each article. This implies that the snapshot file name
`snaps.md` - this is impossible to change right now (I have reporter
this feature request separately). We enhanced this a little using
`variant` argument so that it is `{variant}/snaps.md`. An alternative
approach would be to have a separate file for each article. This would
create snapshot file in accordance to the name of the testing file.
Together with `variant` could have `{stable/devel}/table_ada/adat01.md`.
This would be much more convenient for browsing as well as this would
follow recommendations.
I have been thinking about this and decided to keep it as is. The main
rationale would be to make it super easy to include article in the
regression testing. Currently it's enough to change the article code.
The recommended way would also include additional step of adding a new
test file - something that I would like to avoid.
On the other hand, the structure is pretty much stable and I personally
don't foresee any significant changes. Therefore I think it should be ok
to have separate test file for each article. Also, it's not a big deal
if it will be missing - no test file means no test so nothing will fail.
If done properly, this could also unblock `devtools::test(filter =
"adat01")` type of things - however I am not sure if devs are using
this. I'm interested in hearing your thoughts on this.

---------

Co-authored-by: Emily de la Rua <59304861+edelarua@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>